Add disable_anti_spoof option to bridges
Although the "allowed_address_pairs" neutron extension supports the idea
of completely disabling anti-spoofing, there are cases where users want
to disable anti-spoofing but are not using "allowed_address_pairs". This
change supports this use case.

A new field is added to bridges, "disable_anti_spoof", that is configurable
by the user and has the following behavior:

* If disable_anti_spoof is set to false, then new neutron ports created on this
  bridge will have anti spoof rules as normal.
* If disable_anti_spoof is set to true, then new neutron ports created on this
  bridge will not have any anti spoof rules.

NOTES
* Changing the value of disable_anti_spoof on a bridge will not affect ports
  that already exist on this bridge.
* The default value is false, so the behavior will not change for users not
  using this config.
* Although this affects ports on a neutron network, this is not configurable
  through neutron. You have to use the midonet api (midonet-cli uses the
  midonet-api).

+++ midolman/src/main/java/org/midonet/cluster/data/neutron/SecurityGroupZkManager.java
@@ -36,6 +36,7 @@ import org.midonet.midolman.state.PathBuilder;
 import org.midonet.midolman.state.PortConfig;
 import org.midonet.midolman.state.StateAccessException;
 import org.midonet.midolman.state.ZkManager;
+import org.midonet.midolman.state.zkManagers.BridgeZkManager;
 import org.midonet.midolman.state.zkManagers.ChainZkManager;
 import org.midonet.midolman.state.zkManagers.ChainZkManager.ChainConfig;
 import org.midonet.midolman.state.zkManagers.IpAddrGroupZkManager;
@@ -51,11 +52,13 @@ public class SecurityGroupZkManager extends BaseZkManager {
     private final NetworkZkManager networkZkManager;
     private final PortZkManager portZkManager;
     private final RuleZkManager ruleZkManager;
+    private final BridgeZkManager bridgeZkManager;
 
     @Inject
     public SecurityGroupZkManager(ZkManager zk,
                                   PathBuilder paths,
                                   Serializer serializer,
+                                  BridgeZkManager bridgeZkManager,
                                   ChainZkManager chainZkManager,
                                   IpAddrGroupZkManager ipAddrGroupZkManager,
                                   NetworkZkManager networkZkManager,
@@ -67,6 +70,7 @@ public class SecurityGroupZkManager extends BaseZkManager {
         this.networkZkManager = networkZkManager;
         this.portZkManager = portZkManager;
         this.ruleZkManager = ruleZkManager;
+        this.bridgeZkManager = bridgeZkManager;
     }
 
     private void prepareCreateChain(List<Op> ops, UUID chainId, String name,
@@ -126,25 +130,28 @@ public class SecurityGroupZkManager extends BaseZkManager {
                          .isReturnFlow()
                          .accept());
 
-        // IP spoofing protection for in_chain
-        for (IPAllocation fixedIp : port.fixedIps) {
+        BridgeZkManager.BridgeConfig bridgeConfig = bridgeZkManager.get(port.networkId);
+        if (!bridgeConfig.disableAntiSpoof) {
+            // IP spoofing protection for in_chain
+            for (IPAllocation fixedIp : port.fixedIps) {
 
-            Subnet subnet = networkZkManager.getSubnet(fixedIp.subnetId);
-            IPSubnet ipSub = subnet.isIpv4()
-                             ? fixedIp.ipv4Subnet() : fixedIp.ipv6Subnet();
-            Rule ipSpoofProtectionRule = new RuleBuilder(inChainId)
-                .isAnyFragmentState()
-                .notFromSubnet(ipSub)
-                .drop();
-            inRules.add(ipSpoofProtectionRule);
+                Subnet subnet = networkZkManager.getSubnet(fixedIp.subnetId);
+                IPSubnet ipSub = subnet.isIpv4()
+                        ? fixedIp.ipv4Subnet() : fixedIp.ipv6Subnet();
+                Rule ipSpoofProtectionRule = new RuleBuilder(inChainId)
+                        .isAnyFragmentState()
+                        .notFromSubnet(ipSub)
+                        .drop();
+                inRules.add(ipSpoofProtectionRule);
+            }
+
+            // MAC spoofing protection for in_chain
+            Rule macSpoofProtectionRule = new RuleBuilder(inChainId)
+                    .notFromMac(port.macAddress())
+                    .isAnyFragmentState()
+                    .drop();
+            inRules.add(macSpoofProtectionRule);
         }
-
-        // MAC spoofing protection for in_chain
-        Rule macSpoofProtectionRule = new RuleBuilder(inChainId)
-            .notFromMac(port.macAddress())
-            .isAnyFragmentState()
-            .drop();
-        inRules.add(macSpoofProtectionRule);
 
         // Add reverse flow matching for in_chain.
         inRules.add(new RuleBuilder(inChainId)
